[fix] Partner페이지 확장자 수정, Component 수정
